+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [core] remove lasting references to ec classes
 * [core] add meta block to block robots indexing for some pages
+* [booking] add to scheduling option to force usage of packages to an area
 
 ## 2.5.1