+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changes
 
+## 2.6.0
+
+* [booking_settings] add getDefault to BkNightWE class 
+* [core] remove lasting references to ec classes
+* [core] add meta block to block robots indexing for some pages
+* [core] on user join request, request a comment from user to be added to email
+* [booking] add to scheduling option to force usage of packages to an area
+* [booking] hide nightwe entries if pricing does not exists anymore
+* [helpdesk] do not send ticket reply if mail coming from helpdesk
+* [mail] add support for custom headers
+
+## 2.5.1
+
+* [space] fix pending accounts bug
+* [booking] on calendar entry display, allow no customer/no pricing (catch errors on computeDuration)
+* [core][booking] layout fixes on booking edition (periodic)
+
 ## 2.5.0
 
 * [booking_quantities] avoid having two qtes with the same name for the same resource 
@@ -14,6 +31,7 @@
 * [booking] add to calendar a summary view option
 * [core][ldap] do not automatically add ldap users to spaces
 * [core] add option to tables to view all elements
+* [booking] list blocked resources and add a reason column on bkcalentry (booking, holiday, maintenance)
 * [booking] change booking compute (day/night/we/closed) and get details #565
 * [booking] fix package invoicing #606
 * [core] in core user edit, show user spaces and pending spaces #607